requirements.txt
streamlit==1.37.0
pandas==2.2.2
sqlalchemy==2.0.32
python-dotenv==1.0.1
langchain==0.2.11
langchain-openai==0.1.17
plotly==5.22.0

snowflake-connector-python==3.12.1
snowflake-sqlalchemy==1.6.1
